Split out code to separate classes
